lambdas/requirements.txt
boto3==1.21.22
botocore==1.24.22
jmespath==1.0.0
Pillow==9.0.1
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
